repackage Qt installer script code.
add support for installs on windows, which have a unique page.

The separation of the bash shell code and the Qt Installer script
code allows the later to be used in environments were bash is not
available.  This was implemented with key,value pairs passed on
the command line, but it could have been done with environmental
variables using the installer environmentVariable method instead
of the value method.

add Qt 5.12.5 to regression.